+++ utilities/sdf_primitives.py
@@ -102,10 +102,10 @@ def test():
 	num_points = 2
 	
 	points = Uniform(-0.5, 0.5).sample((batch_size, num_points, 3))
-	translations = torch.tensor([0,0.2,0], dtype=float).repeat(batch_size,1)
-	rotations = torch.tensor([0.924,0,0,0.383], dtype=float).repeat(batch_size,1)
-	scales = torch.tensor([0.2,0.5,0.7], dtype=float).repeat(batch_size,1)
-	roundness = torch.tensor([1], dtype=float).repeat(batch_size,1)
+	translations = torch.tensor([0,0.2,0], dtype=torch.float).repeat(batch_size,1)
+	rotations = torch.tensor([0.924,0,0,0.383], dtype=torch.float).repeat(batch_size,1)
+	scales = torch.tensor([0.2,0.5,0.7], dtype=torch.float).repeat(batch_size,1)
+	roundness = torch.tensor([1], dtype=torch.float).repeat(batch_size,1)
 
 	distances = sdf_ellipsoid(points, translations, rotations, scales)
 	print('Sphere SDF Samples:')
